pyproject.toml
[tool.poetry]
name = "prolaio-ds-demo"
version = "0.1.0"
description = ""
authors = ["Vladimir Pasquier <1112241+vpasquier@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
metaflow = "^2.11.1"
google-cloud-storage = "^2.14.0"
google-auth = "^2.27.0"
botocore = "^1.34.41"
kubernetes = "^29.0.0"
pandas = "^2.2.0"
google-cloud = "^0.34.0"
scikit-learn = "^1.4.0"
google-cloud-bigquery = "^3.17.1"
db-dtypes = "^1.2.0"
matplotlib = "^3.8.2"
wandb = "^0.16.3"
fastcore = "^1.5.29"
torch = "^2.2.0"
debugpy = "^1.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
